[BST-7] feat: implement to update groups
